Temporary fix the code-setup url
It is temporary fix, because the path to `aiidalab_widgets_base` app is
not absolute. It will work only if one tries to access the code setup
from a notebook located in the top level of an app folder.
Once the notebook is moved deeper - it won't work anymore.

The issue to get such a Jupyter root folder is
created https://github.com/aiidalab/aiidalab-widgets-base/issues/19