Introduce the concept of the ProcessFollower (#33)
Add the `ProcessFollower` class that is able to follow a process until it is completed. At the
creation time the `ProcessFollower` accepts a list of followers that can check a particular
outcome of the running process and display it. The followers are specified at the moment of
`ProcessFollower` creation. For the moment we have two followers: `ProgressBarWidget` that
displays the progress and `RunningCalcJobOutputWidget` that displays the output of a running
process.