config ssh by providing private key
In setting the remote computer, can firstly connect to the remote
computer with private key provided. Then the password-less connection is
setup. If the private key file name is already in `~/.ssh/` a shortuuid
will be append to distinguish it from the exist one.
Also, the ssh port is explicitly setable with default value 22.

And when setting computer configuration ssh, port number and private key
filename is read from ~/.ssh/config if is explicitly set in the config
file.

Also add configuration of use_login_shell