Add a possibility to config ssh by providing private key, make SSH port configurable (#110).
This PR adds a possibility to put a private SSH key to the `.ssh` folder of AiiDAlab
when the corresponding public key is already present on the remote machine. To
distinguish the existing private key, the new key's file name is appended with a
`shortuuid`

Additionally, this PR makes the PORT setting configurable.  When setting up the
SSH connection, port number and private key filename are read from  `~/.ssh/config`
if they are explicitly set in the config file.