Revise 'reset' function to be app-wide.
Resetting individual steps does not work intuitively and is very hard to
get right. Therefore, clicking on the reset button will reset the whole
app.

Whether the app can be reset is determined by checking the
'can_reset()' function of each step, which in the default implementation
is True whenever a "reset()" function is implemented. Steps can
implement special logic to indicate that they cannot be reset, e.g.,
because a computation is running or so.

The pizza-order example was adjusted to be "resettable".