fix: dependency conflicts
New releases of jupyter-client and nbconvert had conflicting
dependencies:

jupyter-client 7.0.2 requires nest-asyncio>=1.5, but you'll have nest-asyncio 1.4.3 which is incompatible.
nbconvert 6.1.0 requires traitlets>=5.0, but you'll have traitlets 4.3.3 which is incompatible.

The first one conflicts with plumpy, the second conflicts with the
aiidalab package.

Until these are fixed, we limit the versions of jupyter-client and
nbconvert.