NodesTreeWidget: _find_outputs parse nested namespace outputs (#299)
Fixes #262

Previous PR #269 only works for when the outputs has a single layer of
nested namespace. We store the `parent_pk` in order to traverse nodes
level by level, then only namespace is not enough to access for the
outputs in deep level. To solve this, namespaces is a tuple to store the
path so that all the nodes can be visited.

Co-authored-by: Carl Simon Adorf <simon.adorf@epfl.ch>