FEATURE: `WorkChainNodeViewerWidget` (#335)
This PR implements a simple `WorkChainNodeViewerWidget` that prints out the
output of the `verdi process report` command. The dates were skipped to make
output less noisy. Also, only the output reports from the selected work chain is
printed, not from its descendants, as it could be significant for large workflows.