Parse symmetry info and show in cell tab (#328)
Add the symmetry information of 3-D structure to the 'cell' tab of `StructureDataViewer`. 
Also, add a feature that when the atom is selected from the ngview, the tab will automatically changed to 'selection' tab.
Meanwhile, the unit for the cell length and atom positions the 'angstrom' is added to the tab panel.