SmilesWidget: Make openbabel dependency optional
OpenBabel is a fallback to RDKit in this widget,
and as such is not critical to its function.
Since openbabel is no longer automatically installed in the new
aiidalab/full-stack image, it's better to make it optional.

(RDKit also is not part of the new image and needs to be installed)