Test package using pytest-docker with latest full-stack image
The CI test was run by aiidalab-test-app-action which is outdated and
not going to be maintained anymore. Since we have the new
aiidalab-docker-stack image to launch the AiiDAlab instance, we use it
here to run the test by using the pytest-docker and implement selenium
tests directly and explicitly by checking the elements after the
notebooks are opened and loaded.

The new way of testing make the CI tests more robust and easy to write.
It also support further tests on benchmark tests on widgets loading.

- Remove all support of 3.7 since aiida-core not support it