Install firefox/geckodriver in GH workflow for test (#397)
The test failed since somehow the github server doesn't install firefox by default but only the chrome engine. I think the change happened last week both QeApp and AWB encountered the same issue. However, I cannot find the change log on the web.
Install the geckodriver and firefox, and then specify the path to the firefox engine in the selenium test to solve the issue

The test through selenium is still not very robust, it depends on whether the page is successfully loaded, where I set the sleep time to wait to load but not always reach the end. 
The timeout is increased to 10 secs to make sure the test pass.
The aiidalab dependencies add back to allow the doc build pass.