Take screenshots in notebook tests and upload them as artifacts (#411)
The selenium test not only can running operation and check the element from the browser engine, it can also make the screenshots of the web page. It is very useful to have a intuitive look of every notebook test.
The permission of the package inside the container is changed with adding the writing permission for all user, to be able to installed by pip inside the test container.