Add screenshot teardown fixture - approach 1 (#421)
We want to take a screenshot at the end of each selenium test, even if (and especially) when the test fails.
So essentially we need to take the screenshot as part of the test teardown.

We take the approach of having a separate yield fixture `final_screenshot`. It will take the screenshot at the end of test and store to the file named from the test function by stripping the `test_` prefix. Even if the selenium test failed in the middle, it will take the screenshot as well to show from GUI what exactly happened.