Correctly handle supercell selection (#371)
When dealing with the displayed structure, a user often interacts with the supercell, but not the unit cell.
As the result, clicking on the "virtual" atoms leads to an error, since the selected index is out of bounds
w.r.t. to the real structure. This PR fixes the issue and introduces the correct handling of the selection. 

Additionally, we also add automatic tests to make sure the selection works as expected.

Co-authored-by: Aliaksandr Yakutovich <yakutovicha@gmail.com>