Support the inputs namespace for process inputs check widget (#435)
The namespace input is an `AttributDict` which is not supported by the `ProcessInputWidget` since it tries to show the `pk` of the input nodes.
To fix this, in this commit we support the namespace also for the input ports. The entity of inputs is shown in the dropdown list as <namespace>.<embed_port> if it is nested in the namespace. This is implemented by iteration of the input ports and get the keys for all inputs with nested inputs as well.

----------------------------
Co-authored-by: Aliaksandr Yakutovich <yakutovicha@gmail.com>