Remove spinning Thread from WizzardAppWidgetStep
WizzardAppWidget uses an infinite background thread
to continously update titles of the Accordion widget.
It looks like the only reason for this is to implement
a dynamic "spinner".
This thread just needlesly consumes server resources for
a very little benefit. Implementing a "spinner" in this way is silly.

More importantly, it looks like this background thread
is the main reason for the memory leak we're seeing, see
https://github.com/aiidalab/issues/issues/13

Here I remove the thread in favour of a static unicode character,
in line with all the other icons. It would be great if we could use
some dynamic icons, e.g. fa-spinner as used in the SmilesWidget,
but it looks like the Accordion.set_title() function does not parse
HTML.