ProcessMonitor: Increase default timeout (#480)
ProcessMonitor has a background thread that polls the process status. The default interval is now 0.1 seconds which is quite aggressive, and may potentially lead to a memory leak as discussed on https://github.com/aiidalab/issues/issue/13

Here I propose to increase the default to 1.0s since the typical AiiDA workflow will anyway take quite a bit longer. In any case, widgets can override this by passing the timeout parameter to the constructor.