Remove spinning Thread from WizzardAppWidgetStep (#479)
WizzardAppWidget uses an infinite background thread to continuously update the titles of the Accordion widget. It looks like the only reason for this is to implement a dynamic "spinner". This thread needlessly consumes server resources for little benefit. 

More importantly, it looks like this background thread is the main reason for the memory leak we're seeing, see
https://github.com/aiidalab/issues/issues/13.

Here the thread is removed in favour of a static Unicode character, in line with all the other icons. Additionally, more tests were added to the widget to check its behaviour.

---------
Co-authored-by: Aliaksandr Yakutovich <yakutovicha@gmail.com>