Bump scikit-learn
scikit-learn build for version 0.24 started failing
due to new version of Cython so we bump the version to 1.0.
Per docs, this version does not bring any breaking changes.
https://scikit-learn.org/1.0/auto_examples/release_highlights/plot_release_highlights_1_0_0.html

We also only use the sklearn.PCA in SMILES widget,
I've verified that it still works and checked the release notes
that there are not API changes to this method.
https://scikit-learn.org/1.0/whats_new/v1.0.html#changes-1-0