Remove leftover load profile (#481)
The `load_profile()` in `__init__.py` to load the profile in module scope was introduced by PR #12. It is not needed anymore since the computer and code widgets are updated and can update the dropdown dynamically.

More problematically that this makes the profile loaded in the unit tests and overrides the test profile which makes it has to put the module load inside every test. There is also the problem that if the user want to test the widgets in a newly create python environment that doesn't have an AiiDA profile setup impossible.

In this commit, we detect if the profile is loaded by the notebook, if not, we load it for the user and raising a deprecated warning. This makes it possible to keep backward compatibility thus not breaking the app that doesn't load profile manually.