Not create cell for 0D
`_validate_and_fix_ase_cell` function is not needed anymore, since it add a
dummy cell to gas phase molecule. In the old AiiDA, there needed to be default
cell, otherwise StructureData would not work.
But that has been fixed so we should get rid of this login in AWB.

The fix in aiida-core was done here aiidateam/aiida-core#5341
and released in version 2.0 https://github.com/aiidateam/aiida-core/blob/main/CHANGELOG.md#v200---2022-04-27