Allow set key_filename/key_policy in computer configure (#537)
There were some values for authinfo setting that hardcoded such as keyfile_name and key_policy. This commit set those from computer-configure dict. 
This is required by the CSCS MFA configure setup. Where the problem was key filename is set in configure to cscs-key but not passed to the AuthInfo.