Fix threading issue when accessing user instance in _get_codes of ComputationalResourceWidgets (#543)
fixes aiidalab/aiidalab-qe#582

The _get_codes method is used in the multi-threading case by QEapp, and causes the issue. The error comes from we use `user.collection.get_default` which will return the same object which is not thread-safe. The workaround is to get the user by email (which is the unique index for user) with using `user.collectio.get(email=<emaill>)` which will return a new object.