Drop support for py38 (#557)
The support for Python versions follows the same timeline as that of numpy, defined in NEP 029. According to that schema, Python 3.8 is no longer support as of April 14 2023 and so official support is also dropped in AWB.
Meanwhile, the current Docker stack based on Jupyter docker image is currently Python 3.9, and it was never actually 3.8, so this is a safe deprecation. It would only affect users that are still on the old old docker stack, and those are suggested to migrate to use the new aiidalab docker image.