Add test to latest and oldest supported aiida-core (#558)
The package used with aiida-core as base, so aiida-core is the very base dependency that should carefully test to not break the compatibility. We want to support a wide range of `aiida-core` version. 
The test CI added in this commit run both the unit test and the integration test with latest/oldest aiida-core version in the environment, to make sure the backward compatibility.