Adding support for `orm.Containerized` codes (#617)
This is obtained using the `AiidaCodeSetup().code_setup` attribute.

I am not sure is the better way, as we should also provide the possibility to change the image_name and the engine_command via the GUI. However, I think that for now this is acceptable, as the containerized codes are mainly
set up by using the information on the aiida-resource-registry repo. Actually, it is a really new thing and for now only the phonopy@merlin-cpu is defined.