Downgrade minimum traitlets version to 5.4
As we learned in aiidalab/aiidalab-docker-stack#494,
it is really not good if traitlets dependency is installed
outside of the conda enviroment (i.e. to ~/.local) as it may
break the jupyter notebook. Here we downgrade the minimum traitlets
version back to 5.4.

(NOTE: That the original motivation to upgrading to 5.9 was performance,
I believe it is better if we upgrade directly in the image).