Make the setup new codes widget part of the code selection widget optional (#646)
This PR provides a backwards compatible path to make the new code setup widget part of the code selection widget optional.

The PR also adds a flag (true by default) to control if the codes should be pre-fetched. This is useful, for example, in the QE app, where the MVC redesign (#802) provides these codes via models, removing the responsibility from the UI widgets.