Add `None` to empty widget value check (#650)
This PR corrects a guard for unfilled template variables which checks if the value of the corresponding widget is an empty string. Empty widgets may also be represented by `None`, especially dropdown selectors. This PR adds `None` to the check by comparing the widget's value to both, i.e., it replaces `== ""` with `in ("", None)`.