Clarify structure operations (#666)
This PR aims to emphasize the distinction between visualizing the super cell and actually transforming it. It does so by (i) wrapping the viewer in its own accordion titled "View structure", and (ii) adding a comment below the super cell control in the appearance tab of the viewer clarifying usage.